requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dateparser==1.2.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.41.0
numpy==2.2.6
packaging==24.2
pandas==2.3.0
patsy==1.0.1
pillow==11.2.1
plotly==6.1.2
plotly-express==0.4.1
protobuf==6.31.1
pyarrow==20.0.0
PyAudio==0.2.14
pydeck==0.9.1
pydub==0.25.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SpeechRecognition==3.14.3
statsmodels==0.14.4
streamlit==1.45.1
streamlit-audiorecorder==0.0.6
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
watchdog==6.0.0
